+++ src/main/java/com/grepp/matnam/app/controller/api/notification/NotificationApiController.java
@@ -1,15 +1,12 @@
 package com.grepp.matnam.app.controller.api.notification;
 
 import com.grepp.matnam.app.controller.api.notification.payload.NotificationIdsRequest;
-import com.grepp.matnam.app.model.notification.code.NotificationType;
 import com.grepp.matnam.app.model.notification.entity.Notification;
 import com.grepp.matnam.app.model.notification.service.NotificationService;
 import com.grepp.matnam.infra.auth.AuthenticationUtils;
-import com.grepp.matnam.infra.jwt.JwtTokenProvider;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,9 +14,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 @RequiredArgsConstructor
 @Slf4j
@@ -28,7 +23,6 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 public class NotificationApiController {
 
     private final NotificationService notificationService;
-    private final SSEController sseController;
 
     @GetMapping("/unread-count")
     public ResponseEntity<Long> getUnreadCount() {
@@ -68,14 +62,4 @@ public class NotificationApiController {
         return ResponseEntity.ok().build();
     }
 
-    // 테스트용 알림 생성 API
-    @PostMapping("/create")
-    public ResponseEntity<Notification> createTestNotification(@RequestParam String message, @RequestParam String type, @RequestParam(required = false) String link) {
-        String currentUserId = AuthenticationUtils.getCurrentUserId();
-        NotificationType notificationType = NotificationType.valueOf(type.toUpperCase());
-        Notification notification = notificationService.createNotification(currentUserId, notificationType, message, link);
-        sseController.sendNotificationToUser(currentUserId, "newMessage", notification);
-        sseController.sendNotificationToUser(currentUserId, "unreadCount", notificationService.getUnreadNotificationCount(currentUserId));
-        return ResponseEntity.ok(notification);
-    }
 }
