+++ src/main/java/com/grepp/matnam/app/controller/api/notification/payload/BroadcastNotificationRequest.java
@@ -1,0 +1,8 @@
+package com.grepp.matnam.app.controller.api.notification.payload;
+
+import lombok.Data;
+
+@Data
+public class BroadcastNotificationRequest {
+    private String content;
+}
