+++ src/main/resources/templates/admin/fragments/header.html
@@ -33,7 +33,7 @@
           <button class="notification-tab" data-tab="system">시스템</button>
         </div>
         <div class="notification-list">
-          
+
         </div>
         <div class="notification-footer">
           <a th:href="@{/admin/notification}" class="view-all-notifications">모든 알림 보기</a>
@@ -49,7 +49,7 @@
       <span id="profile-name">관리자</span>
     </div>
   </div>
-  <script th:src="@{/js/notification.js}"></script>
+  <script th:src="@{/js/notification-common.js}"></script>
 </header>
 </body>
 </html>
