+++ src/main/resources/templates/admin/notification.html
@@ -7,7 +7,7 @@
     <div layout:fragment="content">
         <h1>알림 관리</h1>
         <div class="action-bar">
-            <button class="primary-btn" id="add-notification-btn"><i class="fas fa-plus"></i> 새 공지 작성</button>
+            <button class="primary-btn" id="create-notification-btn"><i class="fas fa-plus"></i> 새 공지 작성</button>
             <div class="search-filter">
                 <input type="text" placeholder="공지 내용 검색">
                 <button class="search-btn"><i class="fas fa-search"></i></button>
@@ -71,36 +71,32 @@
             <button class="page-btn next" th:disabled="${currentPage == totalPages}"><i class="fas fa-chevron-right"></i></button>
         </div>
         
-        <!-- 알림 추가/수정 모달 -->
-        <div class="modal" id="notificationModal">
+        <!-- 새 공지사항 작성 모달 -->
+        <div id="create-notification-modal" class="modal">
             <div class="modal-content">
                 <div class="modal-header">
-                    <h2 id="notification-modal-title">공지 작성</h2>
+                    <h2>새 공지사항 작성</h2>
                     <span class="close-modal">&times;</span>
                 </div>
                 <div class="modal-body">
-                    <form id="notificationForm">
-                        <input type="hidden" id="notification-id">
+                    <form id="notification-form">
                         <div class="form-group">
-                            <label for="notification-title">제목</label>
-                            <input type="text" id="notification-title" required>
-                        </div>
-                        <div class="form-group">
-                            <label for="notification-content">내용</label>
-                            <textarea id="notification-content" rows="6" required></textarea>
-                        </div>
-                        <div class="form-group">
-                            <label for="notification-status">상태</label>
-                            <select id="notification-status">
-                                <option value="active">활성</option>
-                                <option value="inactive">비활성</option>
+                            <label for="notification-target">알림 대상</label>
+                            <select id="notification-target" name="notification-target" class="form-control">
+                                <option value="all">전체</option>
+                                <option value="specific" disabled>직접 지정 (준비 중)</option>
                             </select>
+                        </div>
+
+                        <div class="form-group">
+                            <label for="notification-content">메시지 내용</label>
+                            <textarea id="notification-content" name="notification-content" rows="5" placeholder="공지사항 메시지 내용을 입력하세요" required></textarea>
                         </div>
                     </form>
                 </div>
                 <div class="modal-footer">
-                    <button class="secondary-btn cancel-btn">취소</button>
-                    <button class="primary-btn save-btn">저장</button>
+                    <button class="btn cancel-btn">취소</button>
+                    <button class="btn primary-btn" id="send-notification-btn">공지사항 발송</button>
                 </div>
             </div>
         </div>
